requirements.txt
hidapi>=0.14.0
requests>=2.31.0
rich>=13.7.0
